feat: moving chat details into profile panel and settings panel standalone
