+++ packages/client/package.json
@@ -28,23 +28,23 @@
 	},
 	"dependencies": {
 		"@versini/auth-provider": "7.4.0",
-		"@versini/ui-bubble": "2.0.11",
-		"@versini/ui-button": "3.2.0",
+		"@versini/ui-bubble": "2.0.12",
+		"@versini/ui-button": "3.2.1",
 		"@versini/ui-card": "2.0.5",
 		"@versini/ui-footer": "2.0.5",
 		"@versini/ui-hooks": "4.3.0",
-		"@versini/ui-icons": "3.0.2",
+		"@versini/ui-icons": "3.1.0",
 		"@versini/ui-main": "2.0.5",
-		"@versini/ui-menu": "2.0.6",
-		"@versini/ui-panel": "1.2.13",
+		"@versini/ui-menu": "2.0.7",
+		"@versini/ui-panel": "1.2.14",
 		"@versini/ui-spinner": "2.0.5",
 		"@versini/ui-system": "2.0.6",
-		"@versini/ui-table": "2.0.8",
+		"@versini/ui-table": "2.0.9",
 		"@versini/ui-textarea": "2.1.1",
 		"@versini/ui-textinput": "2.0.5",
 		"@versini/ui-toggle": "2.0.5",
 		"@versini/ui-togglegroup": "2.0.5",
-		"@versini/ui-truncate": "2.0.8",
+		"@versini/ui-truncate": "2.0.9",
 		"autoprefixer": "10.4.20",
 		"clsx": "2.1.1",
 		"postcss": "8.4.49",
