+++ packages/client/src/modules/Messages/MessagesContainerHeader.tsx
@@ -5,7 +5,7 @@ import { Suspense, lazy } from "react";
 import { APP_MOTTO, APP_NAME } from "../../common/strings";
 
 const LazyHeader = lazy(
-	() => import(/* webpackChunkName: "LazyHeader" */ "./LazyHeader"),
+	() => import(/* webpackChunkName: "LazyHeader" */ "../Header/LazyHeader"),
 );
 
 export const MessagesContainerHeader = () => {
