+++ packages/client/src/modules/Profile/ProfileContent.tsx
@@ -1,58 +1,72 @@
 import { AUTH_TYPES, useAuth } from "@versini/auth-provider";
 import { ButtonIcon } from "@versini/ui-button";
 import { Card } from "@versini/ui-card";
-import { useLocalStorage, useUniqueId } from "@versini/ui-hooks";
-import { IconEdit, IconPasskey } from "@versini/ui-icons";
+import { useUniqueId } from "@versini/ui-hooks";
+import { IconPasskey } from "@versini/ui-icons";
 import { Flexgrid, FlexgridItem } from "@versini/ui-system";
-import { Toggle } from "@versini/ui-toggle";
-import { useState } from "react";
 
+import { useContext } from "react";
 import {
-	LOCAL_STORAGE_CHAT_DETAILS,
-	LOCAL_STORAGE_PREFIX,
+	DEFAULT_AI_ENGINE,
+	GPT4_MAX_TOKENS,
+	ROLE_ASSISTANT,
 } from "../../common/constants";
-import { CARDS } from "../../common/strings";
+import { CARDS, NA } from "../../common/strings";
+import type { MessageProps } from "../../common/types";
+import {
+	durationFormatter,
+	extractAverage,
+	numberFormatter,
+	pluralize,
+	renderDataAsList,
+} from "../../common/utilities";
+import { AppContext } from "../App/AppContext";
 
-import { renderDataAsList } from "../../common/utilities";
-import { FineTuningPanel } from "./FineTuning";
-import { TagsPanel } from "./Tags";
+type ChatStats = {
+	stats: {
+		averageProcessingTimes: number;
+		totalChats: number;
+	};
+};
 
-export const ProfileContent = () => {
+const averageProcessingTimeFormatter = (value: number) =>
+	durationFormatter(value);
+
+const getAverageProcessingTimePerSession = (
+	chatSession?: { message: MessageProps }[],
+) => {
+	if (!chatSession || chatSession.length === 0) {
+		return NA;
+	}
+
+	const processingTimes = chatSession
+		.filter(
+			(message) =>
+				message?.message?.role === ROLE_ASSISTANT &&
+				typeof message?.message?.processingTime === "number",
+		)
+		.map((data) => data.message.processingTime);
+
+	return extractAverage({
+		data: processingTimes,
+		formatter: averageProcessingTimeFormatter,
+	});
+};
+
+export const ProfileContent = ({ stats }: ChatStats) => {
 	const { isAuthenticated, user, registeringForPasskey, authenticationType } =
 		useAuth();
-	const [showEngineDetails, setShowEngineDetails] = useLocalStorage({
-		key: LOCAL_STORAGE_PREFIX + LOCAL_STORAGE_CHAT_DETAILS,
-		initialValue: false,
-	});
-
-	const [showFineTuning, setShowFineTuning] = useState(false);
-	const [showTags, setShowTags] = useState(false);
+	const { state } = useContext(AppContext);
+	const remainingTokens = GPT4_MAX_TOKENS - Number(state?.usage);
 
 	const listId = useUniqueId();
+	const listIdCurrent = useUniqueId();
+	const listIdMain = useUniqueId();
+
 	const endUser = user?.username || "";
-
-	const onToggleEngineDetails = (checked: boolean) => {
-		setShowEngineDetails(checked);
-	};
-
-	const onClickFineTuning = () => {
-		setShowFineTuning(!showFineTuning);
-	};
-
-	const onClickTags = () => {
-		setShowTags(!showTags);
-	};
 
 	return isAuthenticated && endUser ? (
 		<>
-			{showFineTuning && (
-				<FineTuningPanel
-					open={showFineTuning}
-					onOpenChange={setShowFineTuning}
-				/>
-			)}
-			{showTags && <TagsPanel open={showTags} onOpenChange={setShowTags} />}
-
 			<Card
 				header={CARDS.PREFERENCES.TITLE}
 				className="prose-dark dark:prose-lighter"
@@ -60,33 +74,7 @@ export const ProfileContent = () => {
 				{renderDataAsList(listId, {
 					[CARDS.PREFERENCES.NAME]: endUser,
 					[CARDS.PREFERENCES.EMAIL]: user?.email || "",
-					[CARDS.PREFERENCES.ENGINE_DETAILS]: (
-						<Toggle
-							noBorder
-							labelHidden
-							label={CARDS.PREFERENCES.ENGINE_DETAILS}
-							name={CARDS.PREFERENCES.ENGINE_DETAILS}
-							onChange={onToggleEngineDetails}
-							checked={showEngineDetails}
-						/>
-					),
 				})}
-				<ButtonIcon
-					className="mt-2"
-					size="small"
-					onClick={onClickFineTuning}
-					labelLeft="Engine Fine Tuning"
-				>
-					<IconEdit size="size-3" monotone />
-				</ButtonIcon>
-				<ButtonIcon
-					className="mt-2 ml-2"
-					size="small"
-					onClick={onClickTags}
-					labelLeft="Tags"
-				>
-					<IconEdit size="size-3" monotone />
-				</ButtonIcon>
 			</Card>
 
 			{authenticationType !== AUTH_TYPES.PASSKEY && (
@@ -116,6 +104,33 @@ export const ProfileContent = () => {
 					</ButtonIcon>
 				</Card>
 			)}
+
+			<Card
+				header={CARDS.CURRENT_STATISTICS.TITLE}
+				className="prose-dark dark:prose-lighter mt-4"
+			>
+				{renderDataAsList(listIdCurrent, {
+					[CARDS.CURRENT_STATISTICS.MODEL_NAME]:
+						state?.model || DEFAULT_AI_ENGINE,
+					[CARDS.CURRENT_STATISTICS.TOKENS_USED]: state?.usage,
+					[CARDS.CURRENT_STATISTICS.REMAINING_TOKENS]:
+						numberFormatter.format(remainingTokens),
+					[CARDS.CURRENT_STATISTICS.PROCESSING_TIME]:
+						getAverageProcessingTimePerSession(state?.messages),
+				})}
+			</Card>
+			<Card
+				header={CARDS.MAIN_STATISTICS.TITLE}
+				className="prose-dark dark:prose-lighter mt-4"
+			>
+				{renderDataAsList(listIdMain, {
+					[pluralize(CARDS.MAIN_STATISTICS.TOTAL, stats.totalChats)]:
+						stats.totalChats,
+					[CARDS.MAIN_STATISTICS.PROCESSING_TIME]: durationFormatter(
+						stats.averageProcessingTimes,
+					),
+				})}
+			</Card>
 		</>
 	) : null;
 };
